Add user id to all creation and read db

+++ src/services/prisma/basic-data-service.ts
@@ -4,15 +4,19 @@ import { BasicData } from '../../types';
 export class BasicDataService {
 	static prisma = new PrismaClient();
 
-	async getBasicData() {
-		const basicData = await BasicDataService.prisma.basicData.findMany();
+	async getBasicData(userId: string) {
+		const basicData = await BasicDataService.prisma.basicData.findMany({
+			where: {
+				user_id: userId,
+			},
+		});
 		return basicData;
 	}
 
-	async getSpecificBasicDataById(id: string) {
-		const basicData = await BasicDataService.prisma.basicData.findUnique({
+	async getSpecificBasicDataById(id: string, userId: string) {
+		const basicData = await BasicDataService.prisma.basicData.findFirst({
 			where: {
-				id,
+				AND: [{ id: id }, { user_id: userId }],
 			},
 		});
 		return basicData;
@@ -23,7 +27,11 @@ export class BasicDataService {
 		return basicData;
 	}
 
-	async updateBasicData(id: string, data: BasicData) {
+	async updateBasicData(id: string, data: BasicData, userId: string) {
+		if (!this.basicDataExists(id, userId)) {
+			throw new Error('Basic data does not exist');
+		}
+
 		const basicData = await BasicDataService.prisma.basicData.update({
 			where: {
 				id,
@@ -33,7 +41,11 @@ export class BasicDataService {
 		return basicData;
 	}
 
-	async deleteBasicData(id: string) {
+	async deleteBasicData(id: string, userId: string) {
+		if (!this.basicDataExists(id, userId)) {
+			throw new Error('Basic data does not exist');
+		}
+
 		const basicData = await BasicDataService.prisma.basicData.delete({
 			where: {
 				id,
@@ -45,19 +57,27 @@ export class BasicDataService {
 	async getDollarBlueValue() {
 		const dollarBlueValue = await BasicDataService.prisma.basicData.findFirst({
 			where: {
-				name: 'dollar blue',
+				id: '35981cfb-3c76-46a3-85a3-a4c379987582',
 			},
 		});
 		return dollarBlueValue;
 	}
 
-	async updateDollarBlueValue(value: number) {
+	async updateDollarBlueValue(id: string, value: number) {
 		await BasicDataService.prisma.basicData.updateMany({
 			where: {
-				name: 'dollar blue',
+				id: id,
 			},
 			data: {
 				value: value,
+			},
+		});
+	}
+
+	async basicDataExists(id: string, userId: string) {
+		return await BasicDataService.prisma.basicData.findFirst({
+			where: {
+				AND: [{ id: id }, { user_id: userId }],
 			},
 		});
 	}
